Move processes to jobs directory
